refactor: HTTP 요청 처리 전반을 담당하는 클래스 정의
- 의도를 명확히 드러내기 위해 기존 HTTP 요청을 처리하는 HttpProcessor 클래스명 수정
  - 기존: HttpProcessor
  - 변경: HttpRequestProcessor
- HTTP 요청 처리 시 예외 처리 응집을 위해 HTTP 요청 처리 전반을 담당하는 HttpProcessor 정의
